requirements.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
packaging==20.8
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1
pyparsing==2.4.7
pytest==6.2.1
pytest-cov==2.11.1
pytest-django==4.1.0
python-dateutil==2.8.1
coverage==5.3.1


